Typo & moved get place id stuff to real_estate.py
